+++ lib/rubocop/cop/variable_force/assignment.rb
@@ -11,7 +11,7 @@ module RuboCop
         REFERENCE_PENETRABLE_BRANCH_TYPES = %w(rescue_main ensure_main).freeze
 
         attr_reader :node, :variable, :referenced
-        alias_method :referenced?, :referenced
+        alias referenced? referenced
 
         def initialize(node, variable)
           unless VARIABLE_ASSIGNMENT_TYPES.include?(node.type)