+++ lib/rubocop/cop/variable_force/variable.rb
@@ -12,7 +12,7 @@ module RuboCop
 
         attr_reader :name, :declaration_node, :scope,
                     :assignments, :references, :captured_by_block
-        alias_method :captured_by_block?, :captured_by_block
+        alias captured_by_block? captured_by_block
 
         def initialize(name, declaration_node, scope)
           unless VARIABLE_DECLARATION_TYPES.include?(declaration_node.type)